--- a/demos/cython/hello_world/hello.pyx
+++ b/demos/cython/hello_world/hello.pyx
@@ -1,0 +1,2 @@
+def say_hello_to(name):
+    print("Hello %s" % name)