--- a/demos/cython/wrap_cpp/setup.py
+++ b/demos/cython/wrap_cpp/setup.py
@@ -1,0 +1,6 @@
+from distutils.core import setup
+from Cython.Build import cythonize
+
+setup(name="testapp",
+      ext_modules = cythonize('*.pyx'),
+      )