--- a/proto-mpeg/dct_fast.py
+++ b/proto-mpeg/dct_fast.py
@@ -9,3 +9,4 @@ def idct(F):
     f=fft.dct(fft.dct(F,type=3,norm='ortho').T,type=3,norm='ortho').T
     return f
 
+# TODO: Using dct and idct args axis=1 and axis=0 can avoid the need to transpose the arrays. Compare running times?