… demos/cython/dct/build/temp.linux-x86_64-3.5/dct.o
… demos/cython/dct/dct.c
… demos/cython/dct/dct.cpython-35m-x86_64-linux-gnu.so
… demos/cython/hello_world/build/temp.linux-x86_64-3.5/hello.o
… demos/cython/hello_world/hello.c
… demos/cython/hello_world/hello.cpython-35m-x86_64-linux-gnu.so
… demos/cython/wrap_cpp/hello_world.cpp
… proto-mpeg/cython_test.py
… proto-mpeg/dct.py
… proto-mpeg/outputnew.bin
… proto-mpeg/proto_mpeg.pyc
… proto-mpeg/proto_mpeg_x.pyx