No changes to original pure-python code. Created new Cython files proto_mpeg_x.pyx and proto_mpeg_computation.pyx. These are compiled using distutils by running the new file setup.py. The result is an importable library, proto_mpeg_x, that does significantly faster array manipulation. Image_to_blocks function running time improved from 4 seconds to 6ms. Blocks_to_image function running time improved from 2.5 seconds to 13 ms. Cython code can be demoed using cython_test.py.

On my machine, overall running time to encode a single image is 2.7 seconds. Decode is 16.7 seconds.